Create a new Operator class.
Operators are order dependance operation that preform modifications to the particles before the force classes are run. This will be used in the future to dynamically recompute the dust grain charge.

Introduce the first operator that pre-calculates and caches the sub-step modified positions and velocities for use in the forces. This prevents the re-computation of position and velocity in each force.

Convert ShieldedCoulombForce to used these cached values.

In total this yields a ~ 4% speedup on a 1 second 10000 particle plasma crystal.