Convert thermal forces to use a alternating dispatch groups to cache random numbers in a separate serial queue.
This is necessary because random numbers are not thread save. This new code caches the next set of random numbers in a side queue executed asynchronously. All random numbers must be calculated before they can be used so each block is associated with an even or odd group.

This shaves a minute off execution time on a dual core machine because the even or odd group is most likely finished by the time it is needed so no threads need to halt.